Fix caching mechanism on linux
For some reason, caching with fork doesn't seem to work (ie. it seems to
spawn a new config object each time based on the `id()`, such that the
cache is never kept). I don't understand why, but switching to spawn
shouldn't cost performance for long running processes. Plus, it's a
quick and straightforward fix