Speed up MCMC by passing data less often
The concept is presented here:
https://emcee.readthedocs.io/en/stable/tutorials/parallel/#parallel .
We're frequently passing the same data to the log posterior (every time
the model is called, per emcee), which is quite slow. The workaround is
to create a separate log_posterior module, and then initialize each process
with global variables containing the args for calculating the log
posterior. It's not great coding conventions, but it appears to be an
~3x speed up, so it's a good trade off